Created the various swagger based annotaion classes
This is based on the @doctrine projects annotaion classes from the
doctrine-commen package.

Currently not generating json and not ready for show-time. If anyone
looks at this I suggest seeing `tests/SwaggerTests/Fixtures` for a
preview of the upcoming RCI changes that will be coming.

Very open to suggestions and feedback on this redesign of annotations
support.

[no ci]